requirements.txt
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
rapidfuzz>=3.0
fastapi>=0.110
uvicorn[standard]>=0.30
requests>=2.28
anthropic>=0.40
sentence-transformers>=3.0
numpy>=1.26
rank-bm25>=0.2.2
qdrant-client>=1.17
einops>=0.8
psycopg[binary]>=3.1
